+++ spec/features/mailer_spec.rb
@@ -4,6 +4,7 @@ RSpec.describe "Mailer" do
   let(:gemfile) { IO.read("#{project_path}/Gemfile") }
   let(:mailer_config) { IO.read("#{project_path}/config/mailer.rb") }
   let(:dev_config) { IO.read("#{project_path}/config/environments/development.rb") }
+  let(:sidekiq_config) { IO.read("#{project_path}/config/sidekiq.yml") }
 
   before(:all) { drop_dummy_database }
 
@@ -26,6 +27,8 @@ RSpec.describe "Mailer" do
       expect(dev_config).to include("sendgrid_dev_settings = {")
       expect(dev_config).to include("api_key: ENV['SENDGRID_API_KEY']")
     end
+
+    it { expect(sidekiq_config).to include("- mailers") }
   end
 
   context "when selecting aws_ses as mailer" do
@@ -38,5 +41,18 @@ RSpec.describe "Mailer" do
     it { expect(gemfile).to include("letter_opener") }
     it { expect(mailer_config).to include("delivery_method = :aws_sdk") }
     it { expect(dev_config).to include("delivery_method = :letter_opener") }
+    it { expect(sidekiq_config).to include("- mailers") }
+  end
+
+  context "when selecting a mailer and sidekiq" do
+    before :all do
+      drop_dummy_database
+      remove_project_directory
+      create_dummy_project(
+        "background_processor" => true, "email_service" => 'sendgrid'
+      )
+    end
+
+    it { expect(sidekiq_config).to include("- mailers") }
   end
 end