+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## Unreleased
+
+Features:
   - Add Google Tag Manager [#326](https://github.com/platanus/potassium/pull/326)
+  - Update rubocop and rubocop-rspec for potassium and generated projects [#337](https://github.com/platanus/potassium/pull/337)
+
 ## 6.1.0
 
 Features: